+++ src/main/kotlin/blackjack/model/card/Card.kt
@@ -1,4 +1,4 @@
-package blackjack.model
+package blackjack.model.card
 
 class Card(
     val rank: CardRank,
