+++ src/main/kotlin/blackjack/model/card/CardSuit.kt
@@ -1,4 +1,4 @@
-package blackjack.model
+package blackjack.model.card
 
 enum class CardSuit {
     SPADE,
