+++ src/main/kotlin/blackjack/model/game/GameResult.kt
@@ -1,4 +1,6 @@
-package blackjack.model
+package blackjack.model.game
+
+import blackjack.model.rule.WinningResult
 
 class GameResult(
     val dealerResult: Map<WinningResult, Int>,
