+++ src/main/kotlin/blackjack/model/game/UserCommand.kt
@@ -1,4 +1,4 @@
-package blackjack.model
+package blackjack.model.game
 
 enum class UserCommand {
     HIT,
