+++ src/main/kotlin/blackjack/model/participant/Dealer.kt
@@ -1,4 +1,6 @@
-package blackjack.model
+package blackjack.model.participant
+
+import blackjack.model.card.Card
 
 class Dealer(
     name: String = "",
