+++ src/main/kotlin/blackjack/model/participant/Player.kt
@@ -1,4 +1,6 @@
-package blackjack.model
+package blackjack.model.participant
+
+import blackjack.model.card.Card
 
 class Player(
     name: String,
