+++ src/main/kotlin/blackjack/model/rule/WinningResult.kt
@@ -1,4 +1,4 @@
-package blackjack.model
+package blackjack.model.rule
 
 enum class WinningResult {
     WIN,
