+++ src/test/kotlin/blackjack/domain/CardDeckTest.kt
@@ -1,7 +1,7 @@
 package blackjack.domain
 
-import blackjack.model.Card
-import blackjack.model.CardDeck
+import blackjack.model.card.Card
+import blackjack.model.card.CardDeck
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
